Move getRegistrationFields function to registration utils
