+++ src/domain/registrations/types.ts
@@ -31,18 +31,3 @@ export type ReturnParams = {
 export type RegistrationsLocationState = {
   registrationId: string;
 };
-
-export type RegistrationFields = {
-  id: string;
-  atId: string;
-  createdBy: string;
-  currentAttendeeCount: number;
-  currentWaitingListCount: number;
-  enrolmentEndTime: Date | null;
-  enrolmentStartTime: Date | null;
-  event: string;
-  lastModifiedAt: Date | null;
-  maximumAttendeeCapacity: number;
-  registrationUrl: string;
-  waitingListCapacity: number;
-};
