+++ src/domain/registrations/utils.ts
@@ -1,14 +1,12 @@
 import { ApolloClient, NormalizedCacheObject } from '@apollo/client';
 import capitalize from 'lodash/capitalize';
 
-import { ROUTES } from '../../constants';
 import {
   EventTypeId,
-  RegistrationFieldsFragment,
   RegistrationQueryVariables,
   RegistrationsQueryVariables,
 } from '../../generated/graphql';
-import { Language, PathBuilderProps } from '../../types';
+import { PathBuilderProps } from '../../types';
 import addParamsToQueryString from '../../utils/addParamsToQueryString';
 import getPathBuilder from '../../utils/getPathBuilder';
 import queryBuilder from '../../utils/queryBuilder';
@@ -24,7 +22,6 @@ import {
   REGISTRATIONS_PAGE_SIZE,
 } from './constants';
 import {
-  RegistrationFields,
   RegistrationSearchInitialValues,
   RegistrationSearchParam,
   RegistrationSearchParams,
@@ -121,37 +118,6 @@ export const replaceParamsToRegistrationQueryString = (
     queryParams,
     getRegistrationParamValue
   );
-};
-
-export const getRegistrationFields = (
-  registration: RegistrationFieldsFragment,
-  language: Language
-): RegistrationFields => {
-  const id = registration.id || '';
-
-  return {
-    id,
-    atId: registration.atId || '',
-    createdBy: registration.createdBy ?? '',
-    currentAttendeeCount: registration.currentAttendeeCount ?? 0,
-    currentWaitingListCount: registration.currentWaitingListCount ?? 0,
-    enrolmentEndTime: registration.enrolmentEndTime
-      ? new Date(registration.enrolmentEndTime)
-      : null,
-    enrolmentStartTime: registration.enrolmentStartTime
-      ? new Date(registration.enrolmentStartTime)
-      : null,
-    event: registration.event ?? '',
-    lastModifiedAt: registration.lastModifiedAt
-      ? new Date(registration.lastModifiedAt)
-      : null,
-    maximumAttendeeCapacity: registration.maximumAttendeeCapacity ?? 0,
-    registrationUrl: `/${language}${ROUTES.EDIT_REGISTRATION.replace(
-      ':id',
-      id
-    )}`,
-    waitingListCapacity: registration.waitingListCapacity ?? 0,
-  };
 };
 
 /* istanbul ignore next */
